chore: rename accounts to account
