+++ src/renderer/domains/network/account/model.ts
