+++ src/renderer/domains/network/account/types.ts
